requirements.txt
absl-py==0.9.0
astor==0.8.1
backcall==0.1.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
decorator==4.4.1
gast==0.3.3
google-api-core==1.16.0
googleapis-common-protos==1.51.0
google-auth==1.11.0
google-cloud-core==1.2.0
google-cloud-storage==1.25.0
google-cloud-vision==0.41.0
google-pasta==0.1.8
google-resumable-media==0.5.0
grpcio==1.26.0
h5py==2.10.0
idna==2.8
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.15.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.0.1
Markdown==3.1.1
matplotlib==3.0.3
numpy==1.18.1
pandas==0.25.3
parso==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.2
protobuf==3.11.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.5.2
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.14.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-hub==0.7.0
termcolor==1.1.0
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==0.16.0
wrapt==1.11.2
